chore: make space filter buttons into floating action buttons
